% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_pns.R
\name{dados_pns}
\alias{dados_pns}
\title{Carrega os dados da PNS com desenho amostral e colunas derivadas}
\usage{
dados_pns(ano, vars = c(), global = TRUE)
}
\arguments{
\item{ano}{Ano da PNS (2013 ou 2019).}

\item{vars}{Vetor de nomes de variáveis adicionais a serem carregadas.}

\item{global}{Se TRUE (padrão), os objetos são atribuídos no ambiente global.}
}
\value{
Atribui \code{dados_pns_design} e \code{vars} ao ambiente global (ou retorna em lista, se \code{global = FALSE}).
}
\description{
Esta função baixa e prepara os dados da PNS 2013 ou 2019 com as variáveis solicitadas,
o desenho amostral (survey design), e adiciona variáveis derivadas como faixa etária,
região, IMC categorizado e variáveis renomeadas.
}
