% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_pns_2019.R
\name{dados_pns_2019}
\alias{dados_pns_2019}
\title{Carrega dados da PNS 2019 no ambiente global}
\usage{
dados_pns_2019(vars = c())
}
\arguments{
\item{vars}{Vetor de nomes de variáveis adicionais a serem incluídas.}
}
\value{
Nenhum valor é retornado. Objetos são salvos no ambiente global.
}
\description{
Esta função carrega os dados da PNS 2019 com as variáveis solicitadas
e algumas variáveis fixas obrigatórias para o design da amostra.
}
\details{
Os objetos \code{dados_pns_design} e \code{vars} são salvos no ambiente global.
}
