% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_medias_uf.R
\name{tabela_medias_uf}
\alias{tabela_medias_uf}
\title{Calcula médias por UF da variável de interesse}
\usage{
tabela_medias_uf(variaveis = c())
}
\arguments{
\item{variaveis}{Vetor com nome(s) da(s) variável(is) de interesse.}
}
\value{
Data frame com médias, erro padrão, intervalo de confiança e coeficiente de variação por categoria e UF.
}
\description{
Calcula médias por UF da variável de interesse
}
\examples{
tabela_medias_uf(variaveis = c("P040"))
}
