% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_medias.R
\name{tabela_totais_uf}
\alias{tabela_totais_uf}
\title{Tabela de totais por UF e categoria para variável da PNS}
\usage{
tabela_totais_uf(variaveis = c())
}
\arguments{
\item{variaveis}{Vetor com nomes das variáveis da PNS a serem carregadas. A primeira será usada na tabulação por UF.}
}
\value{
Um data frame em formato wide com totais e medidas de erro por UF e categoria.
}
\description{
Calcula os totais, coeficientes de variação e intervalos de confiança por UF e categoria da primeira variável informada,
com base na PNS 2019.
}
