% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar_codigo.R
\name{buscar_codigo}
\alias{buscar_codigo}
\title{Buscar variáveis no dicionário da PNS}
\usage{
buscar_codigo(palavra, ignore_case = TRUE, ano = NULL)
}
\arguments{
\item{palavra}{Palavra ou expressão regular a ser buscada nas perguntas.}

\item{ignore_case}{Lógico. Se \code{TRUE}, a busca ignora maiúsculas e minúsculas.}

\item{ano}{Opcional. Pode ser 2013, 2019 ou \code{NULL} para ambos.}
}
\value{
Um data frame com as colunas \code{codigo}, \code{pergunta} e \code{Ano} das variáveis que correspondem à busca.
}
\description{
Filtra o dicionário da PNS procurando por uma palavra-chave no texto da pergunta.
}
