% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_pns_2019.R
\name{dados_pns_2019}
\alias{dados_pns_2019}
\title{Carrega os dados da PNS 2019 com desenho amostral}
\usage{
dados_pns_2019(
  vars = c(),
  global = TRUE,
  selecionado = TRUE,
  antropometria = TRUE
)
}
\arguments{
\item{vars}{Vetor de nomes de variáveis adicionais a serem carregadas.}

\item{global}{Se TRUE (padrão), os objetos são atribuídos no ambiente global.}

\item{selecionado}{Se TRUE (padrão), inclui apenas os entrevistados selecionados.}

\item{antropometria}{Se TRUE (padrão), inclui variáveis de antropometria.}
}
\value{
Atribui \code{dados_pns_design} e \code{vars} ao ambiente global (ou retorna em lista, se \code{global = FALSE}).
}
\description{
Esta função baixa e prepara os dados da PNS 2019 com as variáveis solicitadas
e o desenho amostral (survey design). Algumas variáveis obrigatórias para o design
são sempre incluídas automaticamente.
}
