% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela.R
\name{tabela}
\alias{tabela}
\title{Geração de tabela com estimativas da PNS}
\usage{
tabela(
  variaveis = c(),
  filtro = NULL,
  dominio = c("nenhum", "UF", "V0026"),
  metrica = c("total", "media", "prop"),
  agrupar_por = NULL
)
}
\arguments{
\item{variaveis}{Vetor de variáveis de interesse.}

\item{filtro}{Expressão lógica para filtragem dos dados.}

\item{dominio}{Agrupamento adicional ("nenhum", "UF", ou "V0026").}

\item{metrica}{Tipo de métrica: "total", "media" ou "prop".}

\item{agrupar_por}{Variáveis adicionais para cruzamento.}
}
\value{
Um data.frame com estimativas, erro amostral e intervalo de confiança.
}
\description{
Geração de tabela com estimativas da PNS
}
