% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela.R
\name{tabela}
\alias{tabela}
\title{Gera tabela a partir do desenho amostral da PNS}
\usage{
tabela(
  variaveis,
  filtro = NULL,
  dominio = NULL,
  metrica = NULL,
  desagregar = NULL
)
}
\arguments{
\item{variaveis}{Variável ou variáveis a serem analisadas (símbolos com ou sem uso de +).}

\item{filtro}{Expressão lógica para filtrar os dados.}

\item{dominio}{Domínio para desagregação: "nenhum", "UF" ou "V0026".}

\item{metrica}{Tipo de métrica: "total", "media" ou "prop".}

\item{desagregar}{Variável(s) para desagregação adicional (símbolos ou uso de +).}
}
\value{
Um data frame com os resultados da estimativa, intervalo de confiança e coeficiente de variação.
}
\description{
Esta função calcula totais, médias ou proporções de variáveis da PNS,
com opção de filtragem, domínio e agrupamento.
}
