% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_totais.R
\name{tabela_totais}
\alias{tabela_totais}
\title{Tabela com totais e intervalos de confiança para variável da PNS}
\usage{
tabela_totais(variaveis = c())
}
\arguments{
\item{variaveis}{Vetor com nomes das variáveis da PNS a serem carregadas. A primeira será usada na tabulação.}
}
\value{
Um data frame com colunas: Categoria, Total, Coeficiente_de_Variacao, Intervalo_inferior, Intervalo_superior.
}
\description{
Calcula os totais, coeficientes de variação e intervalos de confiança para a primeira variável indicada,
usando o desenho amostral da PNS 2019.
}
